backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary>=2.9.0
Pillow>=10.4.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
python-magic==0.4.27
bleach==6.1.0
